# SPEC/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WPM Backend API",
    "description": "Wealth Portfolio Manager Backend API",
    "version": "0.1.0"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Login",
        "description": "POST endpoint for user authentication.\n\nValidates credentials and returns JWT access token.\n\nArgs:\n    request: LoginRequest containing username and password\n    settings: Application settings\n\nReturns:\n    LoginResponse containing access_token and token_type\n\nRaises:\n    HTTPException: 401 if credentials are invalid",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolio/all": {
      "get": {
        "summary": "Get All Positions Endpoint",
        "description": "GET endpoint to retrieve all portfolio positions with pagination and sorting support.\n\nRequires JWT authentication.\n\nArgs:\n    username: Authenticated username (from token)\n    composite_portfolio: Composite portfolio instance (injected via dependency)\n    price_service: Price service instance (injected via dependency)\n    page: Page number (1-indexed, default: 1)\n    size: Number of items per page (default: 20, max: 100)\n    sort_by: Field to sort by (default: \"ticker\")\n    sort_order: Sort order - \"asc\" or \"desc\" (default: \"asc\")\n\nReturns:\n    PortfolioAllResponse containing paginated list of positions and portfolio totals\n\nRaises:\n    HTTPException: 400 if sort_by field is invalid\n    HTTPException: 500 if portfolio data is not available",
        "operationId": "get_all_positions_endpoint_portfolio_all_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Size"
            },
            "description": "Number of items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Field to sort by",
              "default": "ticker",
              "title": "Sort By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort order: 'asc' or 'desc'",
              "default": "asc",
              "title": "Sort Order"
            },
            "description": "Sort order: 'asc' or 'desc'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAllResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/portfolio/all/performance": {
      "get": {
        "summary": "Get Portfolio Performance Endpoint",
        "description": "GET endpoint to retrieve historical portfolio performance data.\n\nRequires JWT authentication.\n\nArgs:\n    username: Authenticated username (from token)\n    historical_portfolio: Historical composite portfolio instance (injected via dependency)\n    cache_data: Performance cache and cache_end_date (injected via dependency)\n    start_date: Optional start date for performance tracking (ISO format YYYY-MM-DD, defaults to portfolio start_date)\n    end_date: Optional end date for performance tracking (ISO format YYYY-MM-DD, defaults to today)\n    granularity: Granularity level - \"daily\" (default), \"weekly\" (Monday-based), or \"monthly\" (start of month)\n\nReturns:\n    PortfolioPerformanceResponse containing filtered list of PortfolioHistoryPoint objects\n\nRaises:\n    HTTPException: 400 if date format is invalid, date range is invalid, granularity is invalid, or end_date > cache_end_date\n    HTTPException: 422 if granularity value is invalid (handled by FastAPI Query validation)\n    HTTPException: 500 if historical portfolio or performance cache is not available",
        "operationId": "get_portfolio_performance_endpoint_portfolio_all_performance_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date for performance tracking (ISO format YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Start date for performance tracking (ISO format YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date for performance tracking (ISO format YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "End date for performance tracking (ISO format YYYY-MM-DD)"
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(daily|weekly|monthly)$",
              "description": "Granularity of history points: 'daily' (default), 'weekly' (Monday-based), or 'monthly' (start of month)",
              "default": "daily",
              "title": "Granularity"
            },
            "description": "Granularity of history points: 'daily' (default), 'weekly' (Monday-based), or 'monthly' (start of month)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPerformanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username",
            "description": "User's login username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "User's login password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Request model for login endpoint."
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT access token for subsequent API requests"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Type of authentication token",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "LoginResponse",
        "description": "Response model for login endpoint."
      },
...
      "PortfolioAllResponse": {
        "properties": {
          "positions": {
            "$ref": "#/components/schemas/Page_Position_",
            "description": "Paginated list of positions"
          },
          "total_market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Market Value",
            "description": "Total market value across all positions in USD"
          },
          "total_cost_basis": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Cost Basis",
            "description": "Total cost basis across all positions in USD"
          },
          "total_unrealized_gain_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Unrealized Gain Loss",
            "description": "Total unrealized gain or loss across all positions in USD"
          }
        },
        "type": "object",
        "required": [
          "positions",
          "total_cost_basis"
        ],
        "title": "PortfolioAllResponse",
        "description": "Response model for /portfolio/all endpoint with paginated positions and portfolio totals."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/login"
          }
        }
      }
    }
  }
}